requirements.txt
colorama==0.4.6
PyYAML==6.0
requests==2.28.2
